+++ CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.3.0] - 2019-02-14
+### Changed
+- Require Dart SDK 2.1+
 ### Added
 - `android_graphics.Bitmap.fromImage` static method
 - `android_graphics.Bitmap.fromAssetImage` static method