+++ CHANGELOG.md
@@ -6,6 +6,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.2] - 2019-02-15
+### Added
+
 ## [0.3.1] - 2019-02-14
 ### Added
 - `android.bluetooth.BluetoothAdapter` class stub
@@ -159,6 +162,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `android_content.Context.packageName` getter
 - `android_content.Context.packageResourcePath` getter
 
+[0.3.2]:  https://github.com/drydart/flutter_android/compare/0.3.1...0.3.2
 [0.3.1]:  https://github.com/drydart/flutter_android/compare/0.3.0...0.3.1
 [0.3.0]:  https://github.com/drydart/flutter_android/compare/0.2.5...0.3.0
 [0.2.5]:  https://github.com/drydart/flutter_android/compare/0.2.4...0.2.5